+++ configs/core.js
@@ -6,6 +6,7 @@ module.exports = {
     es6: true,
   },
   parserOptions: {
+    // ecmaVersion: 6,
     sourceType: 'module',
     ecmaFeatures: {
       jsx: true,