+++ configs/fantasy.js
@@ -1,3 +1,5 @@
 module.exports = {
-  parser: 'babel-eslint', // Warning: Enables all sorts of experimental stuff such as @decorators and (as of yet) non-standard static class property assignments
+  parserOptions: {
+    ecmaVersion: 9,
+  },
 };