+++ starters/eslintrc.js
@@ -1,4 +1,8 @@
 module.exports = require('hxmstyle')({
   // Place your project-specific additions or overrides here
   // using standard ESLint config syntax...
+
+  parserOptions: {
+    // ecmaVersion: 9, // fantasy
+  },
 });