+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 <!-- Add new lines here. Version number will be decided later -->
 
 - ...
+- **BREAKING** feat: Drop support for the `--fantasy` flag – Use
+  `parserOptions.ecmaVersion` instead. Added example in the .esintrc.js
+  starter.
 - feat: Add optional TypeScript support (BYOTS) – (Bring Your Own TypeScript)
 - feat: Add some `async/await` rules
 - feat: Use the React plugin's native `version: "detect"` feature